% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealutils-package.R
\docType{package}
\name{arealutils-package}
\alias{arealutils}
\alias{arealutils-package}
\title{arealutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Your Name \email{valid@email.com}

}
\keyword{internal}
