% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{besag_aghq}
\alias{besag_aghq}
\title{Fit Besag Small Area Estimation model using \code{aghq}.}
\usage{
besag_aghq(sf, k = 3, ii = NULL)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{k}{Number of quadrature points per hyperparameter dimension.}

\item{ii}{The (zero-indexed) indices of the observations held-out.}
}
\description{
Random effects have an improper conditional autoregressive (ICAR)
distribution with (generalised) precision matrix produced using
the \code{\link{nb_to_precision}} function with input \code{nb},
the neighbourhood structure of \code{sf}.
}
\examples{
besag_aghq(mw)
}
