% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besag.R
\name{besag_inla}
\alias{besag_inla}
\title{Fit Besag Small Area Estimation model using \code{R-INLA}.}
\usage{
besag_inla(sf, verbose = FALSE, cores = parallel::detectCores())
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}

\item{cores}{Number of cores, passed to \code{R-INLA}, defaults to \code{parallel::detectCores()}.}
}
\description{
Random effects have an improper conditional autoregressive (ICAR)
distribution. This is implemented by \code{R-INLA} using the
option \code{model = "besag"}.
}
\examples{
besag_inla(mw, cores = 2)
}
