% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{besag_tmb}
\alias{besag_tmb}
\title{Fit Besag Small Area Estimation model using \code{TMB}.}
\usage{
besag_tmb(sf, its = 1000)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to
\code{nlminb}.}
}
\description{
Random effects have an improper conditional autoregressive (ICAR)
distribution.
}
\examples{
besag_tmb(mw, its = 100)
}
