% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthscale.R
\name{best_average}
\alias{best_average}
\title{Compute length-scale such that averagely distant points have correlation \code{p}.}
\usage{
best_average(D, kernel = matern, p = 0.01, ...)
}
\arguments{
\item{D}{Matrix of distances between points.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{p}{A percentage, defaults to 0.01.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
See Best (1999) "Bayesian models for spatially correlated disease and exposure data".
When the average distance between points is below \code{1e-8} returns value one.
}
