% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border_lengths}
\alias{border_lengths}
\title{Create object containing lengths of borders from \code{sf} object.}
\usage{
border_lengths(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\value{
A list of \code{nrow(sf)} data frames which each have columns:
\itemize{
\item \code{origin} The origin node.
\item \code{perimeter} The perimeter of the origin node.
\item \code{touching} The index of a node having shared border with the origin
node.
\item \code{length} The length of border between \code{origin} and
\code{touching}.
\item \code{weight} The proportion of \code{origin}'s border which is with
\code{touching}.
}
}
\description{
The function \code{\link{border_lengths}} creates a dataframe for each area of
\code{sf} containing information about the shared borders with other areas.
Within this package the primary use is within \code{\link{border_precision}}.
}
\examples{
border_lengths(mw)
}
