% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border_precision}
\alias{border_precision}
\title{Create border ICAR precision matrix from \code{sf}.}
\usage{
border_precision(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\value{
An ICAR precision matrix based upon \code{sf}.
}
\description{
This function creates the a weighted ICAR precision matrix \eqn{Q}.
In \eqn{Q} each entry \eqn{Q_{ij}} is equal to the shared length of
border between the two areas (and so zero if they are not
adjacent). The diagonal elements \eqn{Q_{ii}} equal the total border
of each area.
}
\examples{
border_precision(mw)
}
