% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{bym2_tmb}
\alias{bym2_tmb}
\title{Fit BYM2 Small Area Estimation model using \code{TMB}.}
\usage{
bym2_tmb(sf, its = 1000, ii_mis = NULL)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}

\item{ii_mis}{The (zero-indexed) indices of the observations held-out.}
}
\description{
Fit BYM2 Small Area Estimation model using \code{TMB}.
}
\examples{
bym2_tmb(mw, its = 100)
}
