% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbstan.R
\name{bym2_tmbstan}
\alias{bym2_tmbstan}
\title{Fit BYM2 Small Area Estimation model using \code{tmbstan}.}
\usage{
bym2_tmbstan(
  sf,
  nsim_warm = 100,
  nsim_iter = 1000,
  chains = 4,
  cores = parallel::detectCores(),
  ii_mis = NULL
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{rstan}.}

\item{nsim_iter}{Number of samples, passed to \code{rstan}.}

\item{chains}{Number of chains, each of which gets \code{nsim_warm + nsim_iter} samples, passed to \code{rstan}.}

\item{cores}{Number of cores, passed to \code{rstan}, defaults to \code{parallel::detectCores()}.}

\item{ii_mis}{The (zero-indexed) indices of the observations held-out.}
}
\description{
Fit BYM2 Small Area Estimation model using \code{tmbstan}.
}
\examples{
bym2_tmbstan(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
