% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_methods.R
\name{centroid_distance}
\alias{centroid_distance}
\title{Compute distances between area centroids.}
\usage{
centroid_distance(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\description{
Compute distances between area centroids.
}
\examples{
centroid_distance(mw)
}
