% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ci}
\alias{ci}
\title{Cote d'Ivoire DHS 2012 Survey Data.}
\format{
A simple features data frame with 33 rows and 17 variables.
}
\usage{
ci
}
\description{
Cote d'Ivoire DHS 2012 Survey Data.
}
\keyword{datasets}
