% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbstan.R
\name{ck_tmbstan}
\alias{ck_tmbstan}
\title{Fit Bayesian Centroid MVN Small Area Estimation model using \code{tmbstan}.}
\usage{
ck_tmbstan(
  sf,
  nsim_warm = 100,
  nsim_iter = 1000,
  chains = 4,
  cores = parallel::detectCores()
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{rstan}.}

\item{nsim_iter}{Number of samples, passed to \code{rstan}.}

\item{chains}{Number of chains, each of which gets \code{nsim_warm + nsim_iter} samples, passed to \code{rstan}.}

\item{cores}{Number of cores, passed to \code{rstan}, defaults to \code{parallel::detectCores()}}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyperparameters
are given a prior and learnt.
}
\examples{
ck_tmbstan(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
