% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{constant_aghq}
\alias{constant_aghq}
\title{Fit constant Small Area Estimation model using \code{aghq}.}
\usage{
constant_aghq(sf, k = 3)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{k}{Number of quadrature points per hyperparameter dimension.}
}
\description{
Simply fits a constant (the mean). This is useful as a benchmark for other models.
}
\examples{
constant_aghq(mw)
}
