% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant_stan}
\alias{constant_stan}
\title{Fit constant Small Area Estimation model using \code{rstan}.}
\usage{
constant_stan(
  sf,
  nsim_warm = 100,
  nsim_iter = 1000,
  chains = 4,
  cores = parallel::detectCores()
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{rstan}.}

\item{nsim_iter}{Number of samples, passed to \code{rstan}.}

\item{chains}{Number of chains, each of which gets \code{nsim_warm + nsim_iter} samples, passed to \code{rstan}.}

\item{cores}{Number of cores, passed to \code{rstan}, defaults to \code{parallel::detectCores()}}
}
\description{
Simply fits a constant (the mean). This is useful as a benchmark
for other models.
}
\examples{
constant_stan(mw, cores = 2)
}
