% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{constant_tmb}
\alias{constant_tmb}
\title{Fit constant Small Area Estimation model using \code{TMB}.}
\usage{
constant_tmb(sf, its = 1000, ii_mis = NULL)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}

\item{ii_mis}{The (zero-indexed) indices of the observations held-out.}
}
\description{
Simply fits a constant (the mean). This is useful as a benchmark for other models.
}
\examples{
constant_tmb(mw, its = 100)
}
