% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds.R
\name{create_folds}
\alias{create_folds}
\title{Create cross-validation training data sets.}
\usage{
create_folds(sf, remove_cols = c("y"), type = "LOO")
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{remove_cols}{A vector of named columns which are to have entries
replaced by \code{NA} in the training data sets. Defaults to \code{c("y")}.}

\item{type}{One of \code{"LOO"} or \code{"SLOO"}.}
}
\value{
A list of \code{nrow(sf)} training set lists.
Each training set list contains:
\itemize{
\item \code{data} The training data set with left-out entries.
\item \code{held_out} The indices of all held-out regions.
\item \code{predict_on} The indices of regions to be predicted upon.
}
}
\description{
Creates a list of training datasets each of which has certain entries of
\code{remove_cols} replaced by \code{NA} according to the \code{type} of cross-validation.
If \code{type = "LOO"} one entry of each training dataset is replaced by \code{NA},
if \code{type = "SLOO"} one entry and the entries of the neighbours of the area
it corresponds to are replaced by \code{NA}.
}
\examples{
create_folds(mw, remove_cols = c("y", "est"))
}
