% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_sf_grid}
\alias{create_sf_grid}
\title{Create height by width grid \code{sf} object}
\usage{
create_sf_grid(height, width)
}
\arguments{
\item{height}{Number of grid cells height.}

\item{width}{Number of grid cells width.}
}
\value{
\code{height} by \code{width} \code{sf} object.
}
\description{
Create height by width grid \code{sf} object
}
