% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps.R
\name{crps}
\alias{crps}
\title{Compute an approximation to the continuous ranked probability score.}
\usage{
crps(samples, true_value)
}
\arguments{
\item{samples}{A collection of draws from a probability distribution.}

\item{true_value}{The underlying true value that is being compared to.}
}
\value{
The continuous ranked probability score (CRPS)
}
\description{
Compute an approximation to the continuous ranked probability score.
}
