% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbinom.R
\name{dxbinom}
\alias{dxbinom}
\title{Density of the generalised binomial distribution.}
\usage{
dxbinom(x, size, prob, log = FALSE)
}
\arguments{
\item{x}{A real number less than \code{size}.}

\item{size}{A real number of trials.}

\item{prob}{The probability of success.}

\item{log}{Should the returned probability be on the log scale?}
}
\description{
Density of the generalised binomial distribution.
}
