% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{fck_aghq}
\alias{fck_aghq}
\title{Fit Centroid MVN Small Area Estimation model using \code{aghq}.}
\usage{
fck_aghq(sf, k = 3, its = 1000, kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{k}{Number of quadrature points per hyperparameter dimension.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyper-parameters
are fixed.
}
\examples{
fck_aghq(mw, its = 100)
}
