% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{fck_inla}
\alias{fck_inla}
\title{Fit Centroid MVN Small Area Estimation model using \code{R-INLA}.}
\usage{
fck_inla(
  sf,
  verbose = FALSE,
  kernel = matern,
  cores = parallel::detectCores(),
  ...
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{cores}{Number of cores, passed to \code{R-INLA}, defaults to \code{parallel::detectCores()}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}.
}
\examples{
fck_inla(mw, cores = 2)
}
