% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fck.R
\name{fck_stan}
\alias{fck_stan}
\title{Fit Centroid MVN Small Area Estimation model using \code{rstan}.}
\usage{
fck_stan(
  sf,
  bym2 = FALSE,
  nsim_warm = 100,
  nsim_iter = 1000,
  chains = 4,
  cores = parallel::detectCores(),
  kernel = matern,
  ...
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{bym2}{Logical indicating if the centroid spatial random effects should be convoluted
with unstructured IID noise, defaults to \code{FALSE}.}

\item{nsim_warm}{Number of warmup samples, passed to \code{rstan}.}

\item{nsim_iter}{Number of samples, passed to \code{rstan}.}

\item{chains}{Number of chains, each of which gets \code{nsim_warm + nsim_iter} samples, passed to \code{rstan}.}

\item{cores}{Number of cores, passed to \code{rstan}, defaults to \code{parallel::detectCores()}}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyper-parameters
are fixed.
}
\examples{
fck_stan(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
