% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{fck_tmb}
\alias{fck_tmb}
\title{Fit Centroid MVN Small Area Estimation model using \code{TMB}.}
\usage{
fck_tmb(sf, its = 1000, kernel = matern, ii_mis = NULL, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{ii_mis}{The (zero-indexed) indices of the observations held-out.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyper-parameters
are fixed.
}
\examples{
fck_tmb(mw, its = 100)
}
