% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{fik_tmb}
\alias{fik_tmb}
\title{Fit Integrated MVN Small Area Estimation model using \code{TMB}.}
\usage{
fik_tmb(
  sf,
  its = 1000,
  L = 10,
  type = "random",
  kernel = matern,
  ii = NULL,
  ...
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{type}{The \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{ii}{The (zero-indexed) indices of the observations held-out.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{integrated_covariance}}.
}
\examples{
fik_tmb(mw, its = 100)
}
