% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{get_time}
\alias{get_time}
\alias{get_time.inla}
\alias{get_time.inlax}
\alias{get_time.stanfit}
\title{Report the time taken to fit the model.}
\usage{
get_time(fit, ...)

\method{get_time}{inla}(fit, ...)

\method{get_time}{inlax}(fit, ...)

\method{get_time}{stanfit}(fit, ...)
}
\arguments{
\item{fit}{Fitted model.}

\item{...}{Additional arguments passed to \code{get_time}.}
}
\description{
Report the time taken to fit the model.
}
