% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{iid_aghq}
\alias{iid_aghq}
\title{Fit IID Small Area Estimation model using \code{aghq}.}
\usage{
iid_aghq(sf, k = 3, ii = NULL)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{k}{Number of quadrature points per hyperparameter dimension.}

\item{ii}{The (zero-indexed) indices of the observations held-out.}
}
\description{
Random effects are independent and identically distributed.
}
\examples{
iid_aghq(mw)
}
