% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.R
\name{iid_inla}
\alias{iid_inla}
\title{Fit IID Small Area Estimation model using \code{R-INLA}.}
\usage{
iid_inla(sf, verbose = FALSE, cores = parallel::detectCores())
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}

\item{cores}{Number of cores, passed to \code{R-INLA}, defaults to \code{parallel::detectCores()}.}
}
\description{
Random effects are independent and identically distributed.
}
\examples{
iid_inla(mw, cores = 2)
}
