% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{iid_tmb}
\alias{iid_tmb}
\title{Fit IID Small Area Estimation model using \code{TMB}.}
\usage{
iid_tmb(sf, its = 1000)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}
}
\description{
Random effects are independent and identically distributed.
}
\examples{
iid_tmb(mw, its = 100)
}
