% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ik.R
\name{ik_stan}
\alias{ik_stan}
\title{Fit Bayesian Integrated MVN Small Area Estimation model using \code{rstan}.}
\usage{
ik_stan(
  sf,
  bym2 = FALSE,
  L = 10,
  type = "hexagonal",
  nsim_warm = 100,
  nsim_iter = 1000,
  chains = 4,
  cores = parallel::detectCores()
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{bym2}{Logical indicating if the spatial random effects should be convoluted
with unstructured IID noise, defaults to \code{FALSE}.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{type}{The \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{nsim_warm}{Number of warmup samples, passed to \code{rstan}.}

\item{nsim_iter}{Number of samples, passed to \code{rstan}.}

\item{chains}{Number of chains, each of which gets \code{nsim_warm + nsim_iter} samples, passed to \code{rstan}.}

\item{cores}{Number of cores, passed to \code{rstan}, defaults to \code{parallel::detectCores()}}
}
\description{
Fit Bayesian Integrated MVN Small Area Estimation model using \code{rstan}.
}
\examples{
ik_stan(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
