% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbstan.R
\name{ik_tmbstan}
\alias{ik_tmbstan}
\title{Fit Bayesian Integrated MVN Small Area Estimation model using \code{tmbstan}.}
\usage{
ik_tmbstan(sf, its = 1000, L = 10, type = "hexagonal", ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{type}{The \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{integrated_covariance}}. Kernel hyperparameters
are given a prior and learnt.
}
\examples{
ik_tmbstan(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
