% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_methods.R
\name{integrated_covariance}
\alias{integrated_covariance}
\title{Compute integrated kernel Gram matrix.}
\usage{
integrated_covariance(
  sf,
  L = 10,
  kernel = matern,
  type = "hexagonal",
  l = NULL,
  ...
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{type}{The \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{l}{A lengthscale.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Draws \code{S} samples from each area of \code{sf} and averages \code{kernel} over each pair of draws.
If a lengthscale is not provided then the \code{best_average} method is used.
}
\examples{
integrated_covariance(mw)
}
