% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthscale.R
\name{invgamma_prior}
\alias{invgamma_prior}
\title{Find parameters of the inverse Gamma distribution such that \code{plb} of the probability mass is below \code{lb}
and \code{pub} of the probability mass is above \code{ub}.}
\source{
From \href{https://github.com/paul-buerkner/brms/commit/524e738aeeb82e49c5338839c3e10113763b6de1}{brms}
following \href{Michael Betancourt}{https://betanalpha.github.io/assets/case_studies/gp_part3/part3.html}.
}
\usage{
invgamma_prior(lb, ub, plb = 0.01, pub = 0.01)
}
\arguments{
\item{lb}{Lower bound.}

\item{ub}{Upper bound.}

\item{plb}{Proportion of the mass below \code{lb}, defaults to 0.01.}

\item{pub}{Proportion of the mass above \code{lb}, defaults to 0.01.}
}
\description{
Find parameters of the inverse Gamma distribution such that \code{plb} of the probability mass is below \code{lb}
and \code{pub} of the probability mass is above \code{ub}.
}
