% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_df}
\alias{list_to_df}
\title{Bind the rows of a list together, adding an \code{id} column.}
\usage{
list_to_df(list)
}
\arguments{
\item{list}{A list.}
}
\value{
A dataframe.
}
\description{
Bind the rows of a list together, adding an \code{id} column.
}
