% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{mat_to_nb}
\alias{mat_to_nb}
\title{Wrapper function for \code{spdep::mat2listw}.}
\usage{
mat_to_nb(mat)
}
\arguments{
\item{mat}{An adjacency matrix.}
}
\value{
A neighbourhood list object.
}
\description{
Wrapper function for \code{spdep::mat2listw}.
}
\examples{
adj <- rbind(c(0, 1, 0), c(0, 0, 1), c(0, 0, 0))
adj <- adj + t(adj)
nb <- mat_to_nb(adj)
}
