% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw}
\alias{mw}
\title{Malawi DHS 2015 Survey Data.}
\format{
A simple features data frame with 28 rows and 17 variables.
}
\usage{
mw
}
\description{
Malawi DHS 2015 Survey Data.
}
\keyword{datasets}
