% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{nb_to_graph}
\alias{nb_to_graph}
\title{Create a graph from \code{nb}.}
\source{
From \href{https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R}{code} by Mitzi Morris.
}
\usage{
nb_to_graph(nb)
}
\arguments{
\item{nb}{A neighbourhood list object.}
}
\value{
A list containing
\itemize{
\item \code{n} The number of nodes in the graph
\item \code{n_edges} To-do
\item \code{node1} To-do
\item \code{node2} To-do
}
}
\description{
Create a graph from \code{nb}.
}
\examples{
nb <- sf_to_nb(mw)
nb_to_graph(nb)
}
