% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{nb_to_precision}
\alias{nb_to_precision}
\title{Create ICAR precision matrix from \code{nb}.}
\usage{
nb_to_precision(nb, W = NULL, return_sparse = FALSE, scale_precision = FALSE)
}
\arguments{
\item{nb}{A neighbourhood list object.}

\item{W}{A \code{length(nb)} by \code{length(nb)} symmetric weights matrix with zeros for non-neighbouring areas.}

\item{return_sparse}{Should the matrix returned be of class \code{sparseMatrix}?}

\item{scale_precision}{Should the matrix returned be scaled using \code{scale_gmrf_precision}?}
}
\description{
Create ICAR precision matrix from \code{nb}.
}
\examples{
nb <- sf_to_nb(mw)
nb_to_precision(nb)
}
