% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Produce a \code{ggplot2} visualisation of a (neighbourhood) graph.}
\usage{
plot_graph(sf, add_geography = FALSE)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{add_geography}{Should a map of the geography be plotted alongside the graph?}
}
\value{
A \code{ggplot2} plot.
}
\description{
Produce a \code{ggplot2} visualisation of a (neighbourhood) graph.
}
\examples{
plot_graph(mw, add_geography = TRUE)
}
