% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Produce a \code{ggplot2} visualisation of a matrix.}
\usage{
plot_matrix(M)
}
\arguments{
\item{M}{A matrix.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Produce a \code{ggplot2} visualisation of a matrix.
}
\examples{
M <- matrix(1:9, nrow = 3, ncol = 3)
plot_matrix(M)

plot_matrix(nb_to_precision(sf_to_nb(mw)))
}
