% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile}
\alias{quantile}
\title{Compute an the quantile of an observation within a vector of samples.}
\usage{
quantile(samples, true_value)
}
\arguments{
\item{samples}{A collection of draws from a probability distribution.}

\item{true_value}{The underlying true value that is being compared to.}
}
\value{
The quantile.
}
\description{
Compute an the quantile of an observation within a vector of samples.
}
