% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{riebler_gv}
\alias{riebler_gv}
\title{Compute the Riebler generalised variance of a covariance matrix.}
\usage{
riebler_gv(A)
}
\arguments{
\item{A}{A square matrix.}
}
\description{
Let \eqn{A} be a square matrix, then the Riebler generalised variance
is defined as the geometric mean of the marginal variances, given by
\deqn{\sigma_{\mathrm{GV}}^2(A) = \exp \left( \frac{1}{n} \sum_{i = 1}^n \log A_{ii} \right).}
}
