% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rows_to_list}
\alias{rows_to_list}
\title{Convert the rows of a dataframe to a list.}
\usage{
rows_to_list(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A list.
}
\description{
Convert the rows of a dataframe to a list.
}
