% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_quantile.R
\name{sample_quantile}
\alias{sample_quantile}
\title{Compute an the quantile of an observation within a vector of samples.}
\usage{
sample_quantile(samples, true_value)
}
\arguments{
\item{samples}{A collection of draws from a probability distribution.}

\item{true_value}{The underlying true value that is being compared to.}
}
\value{
The quantile.
}
\description{
Compute an the quantile of an observation within a vector of samples.
}
