% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{sf_to_nb}
\alias{sf_to_nb}
\title{Wrapper function for \code{spdep::poly2nb}.}
\usage{
sf_to_nb(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\value{
A neighbourhood list object.
}
\description{
Wrapper function for \code{spdep::poly2nb}.
}
\examples{
nb <- sf_to_nb(mw)
}
