% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tz}
\alias{tz}
\title{Tanzania AIS 2012 Survey Data.}
\format{
A simple features data frame with 159 rows and 17 variables.
}
\usage{
tz
}
\description{
Tanzania AIS 2012 Survey Data.
}
\keyword{datasets}
