% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-naming.R
\name{update_naming}
\alias{update_naming}
\title{A dictionary to convert from internal names to human readable names, used in the manuscript.}
\usage{
update_naming(df)
}
\arguments{
\item{df}{A dataframe with columns \code{geometry}, \code{sim_model} and \code{inf_model}.}
}
\value{
A dataframe where the entries in those columns have been renamed.
}
\description{
A dictionary to convert from internal names to human readable names, used in the manuscript.
}
