% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\alias{waic.inla}
\alias{waic.stanfit}
\alias{waic.sdreport}
\title{Watanabe–Akaike (Widely-Applicable) information criterion generic.}
\usage{
waic(fit, ...)

\method{waic}{inla}(fit, ...)

\method{waic}{stanfit}(fit, ...)

\method{waic}{sdreport}(fit, ...)
}
\arguments{
\item{fit}{Fitted model.}

\item{...}{Additional arguments passed to \code{waic}.}
}
\description{
Watanabe–Akaike (Widely-Applicable) information criterion generic.
}
