% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{wicar_inla}
\alias{wicar_inla}
\title{Fit Weighted ICAR Small Area Estimation model using \code{R-INLA}.}
\usage{
wicar_inla(sf, R, verbose = FALSE, cores = parallel::detectCores())
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{R}{A sparse precision matrix of class \code{"sparseMatrix"}}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}

\item{cores}{Number of cores, passed to \code{R-INLA}, defaults to \code{parallel::detectCores()}.}
}
\description{
Fit Weighted ICAR Small Area Estimation model using \code{R-INLA}.
}
