% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zw}
\alias{zw}
\title{Zimbabwe DHS 2015 Survey Data.}
\format{
A simple features data frame with 60 rows and 17 variables.
}
\usage{
zw
}
\description{
Zimbabwe DHS 2015 Survey Data.
}
\keyword{datasets}
