% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealutils-package.R
\docType{package}
\name{arealutils-package}
\alias{arealutils}
\alias{arealutils-package}
\title{arealutils: Areal Utility Functions}
\description{
Functions used by areal-comparison analysis.
}
\author{
\strong{Maintainer}: Adam Howes \email{ath19@ic.ac.uk}

}
\keyword{internal}
