% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{besag_tmb}
\alias{besag_tmb}
\title{Fit Besag Small Area Estimation model using \code{TMB}.}
\usage{
besag_tmb(sf, its = 1000, ii = NULL)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}

\item{ii}{The (zero-indexed) indices of the observations held-out.}
}
\description{
Random effects have an improper conditional autoregressive (ICAR)
distribution with (generalised) precision matrix produced using
the \code{\link{nb_to_precision}} function with input \code{nb},
the neighbourhood structure of \code{sf}.
}
\examples{
besag_tmb(mw, its = 100)
}
