% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{bym2_aghq}
\alias{bym2_aghq}
\title{Fit BYM2 Small Area Estimation model using \code{aghq}.}
\usage{
bym2_aghq(sf, k = 3, ii_mis = NULL)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{k}{Number of quadrature points per hyperparameter dimension.}

\item{ii_mis}{The (zero-indexed) indices of the observations held-out.}
}
\description{
Fit BYM2 Small Area Estimation model using \code{aghq}.
}
\examples{
bym2_aghq(mw)
}
