% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{ck_aghq}
\alias{ck_aghq}
\title{Fit Bayesian Centroid MVN Small Area Estimation model using \code{aghq}.}
\usage{
ck_aghq(
  sf,
  bym2 = FALSE,
  nsim_warm = 100,
  nsim_iter = 1000,
  chains = 4,
  cores = parallel::detectCores()
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyperparameters
are given a prior and learnt.
}
\examples{
ck_aghq(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
