% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{ck_tmb}
\alias{ck_tmb}
\title{Fit Bayesian Centroid MVN Small Area Estimation model using \code{TMB}.}
\usage{
ck_tmb(sf, its = 1000)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{its}{Number of iterations in outer loop optimisation, passed to \code{nlminb}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyperparameters
are given a prior and learnt.
}
\examples{
ck_tmb(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
