% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{dic}
\alias{dic}
\alias{dic.inla}
\alias{dic.stanfit}
\title{Deviance information criterion generic.}
\usage{
dic(fit, ...)

\method{dic}{inla}(fit, ...)

\method{dic}{stanfit}(fit, ...)
}
\arguments{
\item{fit}{Fitted model.}

\item{...}{Additional arguments passed to \code{dic}.}
}
\description{
Deviance information criterion generic.
}
