% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{ik_aghq}
\alias{ik_aghq}
\title{Fit Bayesian Integrated MVN Small Area Estimation model using \code{aghq}.}
\usage{
ik_aghq(sf, k = 3, L = 10, type = "random", ii = NULL, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{k}{Number of quadrature points per hyperparameter dimension.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{type}{The \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{ii}{The (zero-indexed) indices of the observations held-out.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{integrated_covariance}}. Kernel hyperparameters
are given a prior and learnt.
}
\examples{
ik_aghq(mw)
}
