% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbstan.R
\name{ik_tmbstan}
\alias{ik_tmbstan}
\title{Fit Bayesian Integrated MVN Small Area Estimation model using \code{tmbstan}.}
\usage{
ik_tmbstan(
  sf,
  nsim_warm = 100,
  nsim_iter = 1000,
  chains = 4,
  cores = parallel::detectCores(),
  L = 10,
  type = "hexagonal",
  ii = NULL,
  ...
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{rstan}.}

\item{nsim_iter}{Number of samples, passed to \code{rstan}.}

\item{chains}{Number of chains, each of which gets \code{nsim_warm + nsim_iter} samples, passed to \code{rstan}.}

\item{cores}{Number of cores, passed to \code{rstan}, defaults to \code{parallel::detectCores()}.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{type}{The \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{ii}{The (zero-indexed) indices of the observations held-out.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{integrated_covariance}}. Kernel hyperparameters
are given a prior and learnt.
}
\examples{
ik_tmbstan(mw, nsim_warm = 0, nsim_iter = 100, cores = 2)
}
